+++ Source/Animatable.swift
@@ -1,0 +1,4 @@
+protocol Animatable {
+  var layer: CALayer { get }
+  var frame: CGRect { get }
+}
